+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/endpoints/NoMetricsMultipleLookupDimensionDataServletSpec.groovy
@@ -1,21 +1,16 @@
 package com.yahoo.bard.webservice.web.endpoints
 
-import com.yahoo.bard.webservice.application.AbstractBinderFactory
-import com.yahoo.bard.webservice.config.SystemConfigProvider
+import com.yahoo.bard.webservice.config.BardFeatureFlag
 import com.yahoo.bard.webservice.data.dimension.BardDimensionField
 import com.yahoo.bard.webservice.data.dimension.DimensionDictionary
 import com.yahoo.bard.webservice.util.GroovyTestUtils
 import com.yahoo.bard.webservice.util.JsonSortStrategy
 
 class NoMetricsMultipleLookupDimensionDataServletSpec extends BaseDataServletComponentSpec {
-    String noMetricsQueryBackup
 
     @Override
     def setup() {
-        Boolean noMetricsQueryBackupBoolean = SystemConfigProvider.getInstance().getBooleanProperty(AbstractBinderFactory.REQUIRE_METRICS_IN_QUERY_KEY)
-        noMetricsQueryBackup = noMetricsQueryBackupBoolean == null ? null : noMetricsQueryBackupBoolean.toString()
-
-        SystemConfigProvider.getInstance().setProperty(AbstractBinderFactory.REQUIRE_METRICS_IN_QUERY_KEY, "false")
+        BardFeatureFlag.REQUIRE_METRICS_QUERY.setOn(false)
 
         DimensionDictionary dimensionStore = jtb.configurationLoader.dimensionDictionary
         dimensionStore.findByApiName("sex").with {
@@ -35,6 +30,10 @@ class NoMetricsMultipleLookupDimensionDataServletSpec extends BaseDataServletCom
         }
     }
 
+    @Override
+    def cleanup() {
+        BardFeatureFlag.REQUIRE_METRICS_QUERY.reset()
+    }
 
     @Override
     Class<?>[] getResourceClasses() {
@@ -173,10 +172,5 @@ class NoMetricsMultipleLookupDimensionDataServletSpec extends BaseDataServletCom
                          "species|id": "species3"
                      }]
         }"""
-    }
-
-    @Override
-    def cleanup() {
-        SystemConfigProvider.getInstance().resetProperty(AbstractBinderFactory.REQUIRE_METRICS_IN_QUERY_KEY, noMetricsQueryBackup)
     }
 }